Rename nit to n_iterations and add Result to init
