+++ pybop/optimisers/scipy_optimisers.py
@@ -1,7 +1,8 @@
 import numpy as np
 from scipy.optimize import differential_evolution, minimize
 
-from pybop.optimisers.base_optimiser import BaseOptimiser, Result
+from pybop import Result
+from pybop.optimisers.base_optimiser import BaseOptimiser
 
 
 class BaseSciPyOptimiser(BaseOptimiser):
@@ -70,12 +71,10 @@ class BaseSciPyOptimiser(BaseOptimiser):
         """
         result = self._run_optimiser()
 
-        self._iterations = result.nit
-
         return Result(
             x=result.x,
             final_cost=self.cost(result.x),
-            nit=result.nit,
+            n_iterations=result.nit,
             scipy_result=result,
         )
